Added Football Association agent tests and fixed some issues regarding null in arguments at agent class

+++ src/main/java/users/FootballAssociationAgent.java
@@ -16,6 +16,8 @@ public class FootballAssociationAgent extends User{
     }
 
     public boolean addLeague(LeagueType type) {
+        if(type == null)
+            return false;
         League league = new League(type);
         try {
             return db.addLeague(this, league);
@@ -26,7 +28,23 @@ public class FootballAssociationAgent extends User{
         return false;
     }
 
+    public boolean removeLeague(League league) {
+        if(league == null)
+            return false;
+        try{
+            return db.deleteLeague(this, league);
+        }
+        catch (Exception e){
+            e.printStackTrace();
+        }
+        return false;
+    }
+
+
+
     public boolean addSeasonToLeague(int leagueID,int year,Season season){
+        if(year <= 0 || season == null)
+            return false;
         try {
             LinkedList<League> leagues = db.getLeagues(this);
             for (League league : leagues) {
@@ -42,6 +60,14 @@ public class FootballAssociationAgent extends User{
     }
 
     public boolean addReferee(String firstName, String lastName, String userName, String password, float salary, RefereeType type){
+        if(firstName == null || lastName == null || userName == null || password == null)
+            return false;
+        if(userName.isEmpty() || password.isEmpty())
+            return false;
+        if(type == null)
+            return false;
+        if(salary <= 0)
+            return false;
         Referee referee = new Referee(new User(firstName, lastName, userName, password), salary, type);
         try {
             return db.addReferee(this, referee);
@@ -53,6 +79,8 @@ public class FootballAssociationAgent extends User{
     }
 
     public boolean deleteReferee(Referee referee){
+        if(referee == null)
+            return false;
         try{
             return db.deleteReferee(this, referee);
         }
@@ -63,6 +91,8 @@ public class FootballAssociationAgent extends User{
     }
 
     public boolean addRefereeToLeagueSeason(League league, Season season, Referee referee){
+        if(league == null || season == null || referee == null)
+            return false;
         try{
             LeagueSeasonController leagueSeasonController = db.getLeagueSeasonController(this, league, season);
             if(leagueSeasonController == null){
@@ -74,5 +104,14 @@ public class FootballAssociationAgent extends User{
             e.printStackTrace();
         }
         return false;
+    }
+    public boolean addLeagueSeasonController(Season s,League l){
+        if(l == null || s == null)
+            return false;
+        try {
+            return db.addLeagueSeasonController(this,l,s);
+        } catch (Exception e) {
+            return false;
+        }
     }
 }
